+++ src/client/source/CMakeLists.txt
@@ -1,0 +1,2 @@
+file(GLOB client_source CONFIGURE_DEPENDS "*.c" "*.cpp")
+target_sources(client PRIVATE ${client_source})
